+++ querydsl-codegen/src/main/java/com/querydsl/codegen/DefaultFiler.java
@@ -13,11 +13,20 @@
  */
 package com.querydsl.codegen;
 
+import javax.annotation.processing.ProcessingEnvironment;
+import javax.lang.model.element.Element;
+import java.io.IOException;
+import java.io.Writer;
+import java.util.Collection;
+
 /**
- * {@code SupertypeSerializer} is a {@link Serializer} for supertypes
+ * creates files with {@link javax.annotation.processing.Filer}.
  *
- * @author tiwe
- *
+ * @author f43nd1r
  */
-public interface SupertypeSerializer extends EntitySerializer {
+public class DefaultFiler implements Filer {
+    @Override
+    public Writer createFile(ProcessingEnvironment processingEnvironment, String classname, Collection<? extends Element> elements) throws IOException {
+        return processingEnvironment.getFiler().createSourceFile(classname, elements.toArray(new Element[0])).openWriter();
+    }
 }
