+++ querydsl-codegen/src/main/java/com/querydsl/codegen/Filer.java
@@ -13,11 +13,17 @@
  */
 package com.querydsl.codegen;
 
+import javax.annotation.processing.ProcessingEnvironment;
+import javax.lang.model.element.Element;
+import java.io.IOException;
+import java.io.Writer;
+import java.util.Collection;
+
 /**
- * {@code SupertypeSerializer} is a {@link Serializer} for supertypes
+ * provides files in annotation processing to write classes to
  *
- * @author tiwe
- *
+ * @author f43nd1r
  */
-public interface SupertypeSerializer extends EntitySerializer {
+public interface Filer {
+    Writer createFile(ProcessingEnvironment processingEnvironment, String classname, Collection<? extends Element> elements) throws IOException;
 }
