+++ querydsl-codegen/src/test/java/com/querydsl/codegen/CustomTypeTest.java
@@ -32,7 +32,7 @@ public class CustomTypeTest {
 
     private final TypeMappings typeMappings = new JavaTypeMappings();
 
-    private final EntitySerializer serializer = new EntitySerializer(typeMappings, Collections.<String>emptySet());
+    private final EntitySerializer serializer = new DefaultEntitySerializer(typeMappings, Collections.<String>emptySet());
 
     private final StringWriter writer = new StringWriter();
