+++ querydsl-kotlin-codegen/src/main/kotlin/com/querydsl/kotlin/codegen/KotlinSuperSerializer.kt
@@ -11,13 +11,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.querydsl.codegen;
+package com.querydsl.kotlin.codegen
 
-/**
- * {@code SupertypeSerializer} is a {@link Serializer} for supertypes
- *
- * @author tiwe
- *
- */
-public interface SupertypeSerializer extends EntitySerializer {
+import com.querydsl.codegen.SupertypeSerializer
+import com.querydsl.codegen.TypeMappings
+import javax.inject.Inject
+import javax.inject.Named
+
+class KotlinSuperSerializer @Inject constructor(mappings: TypeMappings, @Named("keywords") keyword: Collection<String>) : KotlinEntitySerializer(mappings, keyword),
+    SupertypeSerializer {
 }
