+++ querydsl-kotlin-codegen/src/main/kotlin/com/querydsl/kotlin/codegen/KotlinTypeMappings.kt
@@ -11,13 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.querydsl.codegen;
+package com.querydsl.kotlin.codegen
 
-/**
- * {@code SupertypeSerializer} is a {@link Serializer} for supertypes
- *
- * @author tiwe
- *
- */
-public interface SupertypeSerializer extends EntitySerializer {
-}
+import com.querydsl.codegen.JavaTypeMappings
+
+class KotlinTypeMappings : JavaTypeMappings()
