+++ querydsl-kotlin-codegen/src/main/resources/META-INF/services/com.querydsl.codegen.Extension
@@ -1,0 +1,1 @@
+com.querydsl.kotlin.codegen.KotlinAptExtension
