+++ querydsl-kotlin-codegen/src/test/kotlin/Entity.kt
@@ -11,13 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.querydsl.codegen;
+class Entity {
+    var name: String? = null
 
-/**
- * {@code SupertypeSerializer} is a {@link Serializer} for supertypes
- *
- * @author tiwe
- *
- */
-public interface SupertypeSerializer extends EntitySerializer {
+    companion object {
+        fun test(o: Any?): String? {
+            return null
+        }
+    }
 }
